requirements/base.txt
Django>=2.1,<2.2
django-environ==0.4.5
django-extensions==2.1.0
djangorestframework==3.8.2
ipython==6.4.0
pytz==2018.5
whitenoise==4.0
django-cors-headers==3.5.0
djangorestframework-simplejwt==4.4.0
